requirements.txt
apache-airflow==3.0.2
redis
requests
